Fix UTC offset snprintf type mismatch
Adjust `UtcOffset` formatting to use signed `%02d` specifiers and cast hour/minute values to 8-bit signed integers before passing them to `snprintf`. This aligns argument types with the format string and prevents signed/unsigned formatting issues in UTC offset output.

+++ common/include/json/json_format_helpers.h
@@ -44,7 +44,10 @@ inline const char* UtcOffset(int32_t hours, uint32_t minutes, char (&buf)[kOffse
     if (kNegative) {
         hours = -hours;
     }
-    snprintf(buf, sizeof(buf), "%c%02d:%02u", kNegative ? '-' : '+', hours, minutes);
+	snprintf(buf, sizeof(buf), "%c%02d:%02d", 
+	         kNegative ? '-' : '+', 
+	         static_cast<int8_t>(hours), 
+	         static_cast<int8_t>(minutes));
     return buf;
 }
 } // namespace format
